feat(study-room-scraper): ✨ add prisma schema for study spaces
